Added warnings to compiler and fixed warnings

+++ effects.cpp
@@ -1,7 +1,7 @@
 #include "wave.h"
 
 void KarplusStrong(int16_t * Buffer, int SamplesToWrite, int SamplingRate, float ToneHz, int Volume) {
-    int SampleWindowSize = SamplingRate / ToneHz;
+    int SampleWindowSize = (int) (SamplingRate / ToneHz);
 
     int16_t *SampleWindowBuffer = Buffer;
     int SampleIndex = 0;
@@ -13,8 +13,8 @@ void KarplusStrong(int16_t * Buffer, int SamplesToWrite, int SamplingRate, float
 
         float SampleValue = ((float) value / RAND_MAX) * 2.0f - 1.0f;
 
-        *Buffer++ = SampleValue * Volume;
-        *Buffer++ = SampleValue * Volume;
+        *Buffer++ = (int16_t) (SampleValue * Volume);
+        *Buffer++ = (int16_t) (SampleValue * Volume);
     }
 
     for (; SampleIndex < SamplesToWrite; SampleIndex++) {
@@ -22,7 +22,7 @@ void KarplusStrong(int16_t * Buffer, int SamplesToWrite, int SamplingRate, float
         SampleWindowBuffer += 2;
         int16_t SecondFeedbackValue = *SampleWindowBuffer;
 
-        int16_t SampleValue = (float)(FirstFeedbackValue + SecondFeedbackValue) * 0.994f / 2;
+        int16_t SampleValue = (int16_t) ((float)(FirstFeedbackValue + SecondFeedbackValue) * 0.994f / 2.0f);
         *Buffer++ = SampleValue;
         *Buffer++ = SampleValue;
     }
