+++ wave.h
@@ -7,9 +7,9 @@
 #define PI 3.14159265f
 
 
-int SamplesPerSecond = 48000;
-int Channels = 2;
-int BytesPerSample = 2;
+uint32_t SamplesPerSecond = 48000;
+uint8_t Channels = 2;
+uint8_t BytesPerSample = 2;
 int BufferLength = 2;
 int BytesPerSecond = Channels * BytesPerSample * SamplesPerSecond;
 int BufferSize = BufferLength * BytesPerSecond;
@@ -19,7 +19,7 @@ struct SoundContext;
 typedef float (* WaveFn)(float TimeIndex, float Tone);
 
 
-void FillBuffer(int16_t * Buffer, int SamplesToWrite, WaveFn Wave, float ToneHz, int Volume);
+void FillBuffer(int16_t * Buffer, uint32_t SamplesToWrite, WaveFn Wave, float ToneHz, int Volume);
 
 #define WAVE_H
 #endif
